Добавлено подтверждение при удалении пред. направления

+++ ecom/ecom-riams/web/src/webapp/WEB-INF/actions/smo_direction/pre-recorded1.jsp
@@ -723,8 +723,8 @@ ul.listTimes li.first {
   			
   		}
   		function deleteTime(aTime,aDirect) {
-  			
-  			WorkCalendarService.deletePreRecord(
+  			if (confirm("Хотите удалить предварительную запись?")) {
+  				WorkCalendarService.deletePreRecord(
   					aTime,
 					
 			     {
@@ -740,6 +740,7 @@ ul.listTimes li.first {
 							}
 						}
      			 }) ;
+  			}
   		}
   		function step6Finish() {
   			step6() ;
